+++ R/collect_draws.R
@@ -132,9 +132,7 @@
     if (is.null(period)) {
       period <- tt
     } else {
-      if (period > tt | period < 1) {
-        stop("Implausible specification of argument 'period'.")
-      }
+      period <- .check_period(period, tt)
     }
   }