+++ R/create_bvarmodel.R
@@ -492,9 +492,9 @@ create_bvarmodel <- function(data, p = 2,
       
       # Create individual model
       result_i <- list("model" = model_i,
-                       "data" = list("original" = list("endogen" = data,
+                       "data" = list("original" = .drop_null(list("endogen" = data,
                                                        "exogen" = exogen,
-                                                       "deterministic" = det_data),
+                                                       "deterministic" = det_data)),
                                      "train" = list("y" = y,
                                                     "x" = x,
                                                     "z" = z)))